Moved exportGraph to ExportModal class
