+++ src/ExportModal.ts
@@ -2,7 +2,6 @@ import { Board } from "jsxgraph";
 import ObsidianGraphs from "main";
 import {  debounce, Modal, normalizePath, Notice, Setting } from "obsidian";
 import { LocationSuggester } from "./LocationSuggester";
-import { Utils } from "./utils";
 
 export class ExportModal extends Modal {
 	plugin: ObsidianGraphs;
@@ -12,7 +11,6 @@ export class ExportModal extends Modal {
 	transparentBackground: boolean;
 	graphs: HTMLElement[];
 	boards: Board[];
-	utils: Utils;
 
 	constructor(plugin: ObsidianGraphs, boards: Board[]) {
 		super(plugin.app);
@@ -22,8 +20,32 @@ export class ExportModal extends Modal {
 		this.exportButtons = [];
 		this.graphs = [];
 		this.boards = boards;
-		this.utils = new Utils();
 		this.svgs = [];
+	}
+
+	exportGraph(graph: Board, transparentBackground: boolean) {
+		const text = graph.renderer.dumpToDataURI();
+		const ar = text.split(",");
+		let  decoded =  atob(ar[1]);
+		const re  = RegExp(/var\(\s*(--[\w-]+)\s*\)/g);
+		let matches = re.exec(decoded);
+		while (matches != null) {
+				decoded = decoded.replace(matches[0], document.body.getCssPropertyValue(matches[1]));
+				matches = re.exec(decoded);
+		}
+
+		const insertPosition = decoded.indexOf(">")+1;
+		let beginning = decoded.slice(0, insertPosition);
+		const ending = decoded.slice(insertPosition);
+
+		if (!transparentBackground) {
+			beginning = beginning.replace("style=\"", "style=\"background-color: " + document.body.getCssPropertyValue("--background-secondary") + "; ");
+		}
+
+		const  style = "<style>.JXG_navigation {display: none;}</style>"
+		decoded = beginning + style + ending;
+
+		return decoded;
 	}
 
 	async onOpen() {
@@ -100,7 +122,7 @@ export class ExportModal extends Modal {
 			const svgContainer = container.createDiv();
 			svgContainer.empty();
 			svgContainer.addClass("svgContainer");
-			const svg = this.utils.exportGraph(this.boards[i], this.transparentBackground);
+			const svg = this.exportGraph(this.boards[i], this.transparentBackground);
 			svgContainer.innerHTML = svg;
 
 			if (svgContainer.firstElementChild?.getAttribute("width") == "0") {
